i2c: Make i2c advance function match new prototypes

+++ include/mraa_adv_func.h
@@ -64,9 +64,9 @@ typedef struct {
     mraa_result_t (*i2c_set_frequency_replace) (mraa_i2c_context dev, mraa_i2c_mode_t mode);
     mraa_result_t (*i2c_address_replace) (mraa_i2c_context dev, uint8_t addr);
     int (*i2c_read_replace) (mraa_i2c_context dev, uint8_t* data, int length);
-    uint8_t (*i2c_read_byte_replace) (mraa_i2c_context dev);
-    uint8_t (*i2c_read_byte_data_replace) (mraa_i2c_context dev, const uint8_t command);
-    uint16_t (*i2c_read_word_data_replace) (mraa_i2c_context dev, const uint8_t command);
+    int (*i2c_read_byte_replace) (mraa_i2c_context dev);
+    int (*i2c_read_byte_data_replace) (mraa_i2c_context dev, const uint8_t command);
+    int (*i2c_read_word_data_replace) (mraa_i2c_context dev, const uint8_t command);
     int (*i2c_read_bytes_data_replace) (mraa_i2c_context dev, uint8_t command, uint8_t* data, int length);
     mraa_result_t (*i2c_write_replace) (mraa_i2c_context dev, const uint8_t* data, int length);
     mraa_result_t (*i2c_write_byte_replace) (mraa_i2c_context dev, uint8_t data);
